requirements.txt
Click
loguru
opencv-python
pydantic
pyyaml
twitchio
